refactor(fiche de paie): déplace les composants
