+++ site/source/components/simulationExplanation/FicheDePaie/Line.tsx
