+++ site/source/components/simulationExplanation/FicheDePaie/utils.ts
