+++ site/source/pages/simulateurs/sasu/components/FicheDePaie/SalaireBrut.tsx
@@ -3,9 +3,9 @@ import { useTranslation } from 'react-i18next'
 import { Condition } from '@/components/EngineValue/Condition'
 import { H3 } from '@/design-system'
 
-import '@/components/FicheDePaie/FicheDePaie.css'
+import '@/components/simulationExplanation/FicheDePaie/FicheDePaie.css'
 
-import { Line } from '@/components/FicheDePaie/Line'
+import { Line } from '@/components/simulationExplanation/FicheDePaie/Line'
 
 export const SalaireBrut = () => {
 	const { t } = useTranslation()
